pyproject.toml
[tool.poetry]
name = "egobot"
version = "0.2.0"
description = "egobot v2"
authors = ["Lana <lan2828@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.8.1"
build = "1.2.2.post1"
cachecontrol = "0.14.1"
cleo = "2.1.0"
colorama = "0.4.6"
crashtest = "0.4.1"
distlib = "0.3.9"
fastjsonschema = "2.20.0"
filelock = "3.16.1"
h11 = "0.14.0"
installer = "0.7.0"
jaraco-classes = "3.4.0"
more-itertools = "10.5.0"
msgpack = "1.1.0"
oauthlib = "3.2.2"
packaging = "24.2"
pexpect = "4.9.0"
pkginfo = "1.11.2"
platformdirs = "4.3.6"
poetry = "1.8.4"
poetry-core = "1.9.1"
poetry-plugin-export = "1.8.0"
ptyprocess = "0.7.0"
pyproject-hooks = "1.2.0"
pytz-deprecation-shim = "0.1.0.post0"
pywin32-ctypes = "0.2.3"
rapidfuzz = "3.10.1"
requests-toolbelt = "1.0.0"
rsa = "4.9"
shellingham = "1.5.4"
six = "1.16.0"
sqlparse = "0.5.1"
tomlkit = "0.13.2"
trove-classifiers = "2024.10.21.16"
virtualenv = "20.27.1"
httpcore = "^1.0.6"
httpx = "^0.27.2"
python-telegram-bot = "^21.7"
keyring = "^24.3.1"
anyio = "^4.6.2.post1"
apscheduler = "^3.10.4"
cachetools = "^5.5.0"
charset-normalizer = "^3.4.0"
cryptography = "^43.0.3"
future = "^1.0.0"
idna = "^3.10"
pyasn1 = "^0.6.1"
pyasn1-modules = "^0.4.1"
pycparser = "^2.22"
python-dotenv = "^1.0.1"
pytz = "^2024.2"
requests = "^2.32.3"
requests-oauthlib = "^2.0.0"
setuptools = "^75.3.0"
sniffio = "^1.3.1"
tornado = "^6.4.1"
typing-extensions = "^4.12.2"
tzdata = "^2024.2"
tzlocal = "^5.2"
urllib3 = "^2.2.3"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
